backend/requirements.txt
# Idea8 Backend Dependencies
# AI-powered idea generation and validation platform


uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
httpx==0.25.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
redis==5.0.1
aiofiles==23.2.1
Pillow==10.1.0
pypdf2==3.0.1
python-docx==1.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
email-validator==2.1.0
beautifulsoup4==4.12.2
jinja2==3.1.2
alembic
watchdog
tenacity
guardrails-ai
pydantic[email]
dirtyjson
pydantic-ai[full]
fastapi
starlette
greenlet